requirement.txt.txt
torch==1.13.1
transformers==4.28.1
datasets==2.3.2
wandb==0.12.21
flask==2.0.3
sqlite3
pandas==1.4.2
psycopg2-binary==2.9.3
mysql-connector-python==8.0.29
cx_Oracle==8.3.0
pymongo==4.0.2
pyodbc==4.0.32
bitsandbytes==0.33.0
peft==0.3.0
nltk==3.7
peft